Moved everything back into the FileController
