--- app/Http/Controllers/ChunkFileUploadController.php
@@ -1,82 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Http\Requests\FileUploadRequest;
-use App\Models\File;
-use App\Rules\Extension;
-use Illuminate\Http\UploadedFile;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Validator;
-use Pion\Laravel\ChunkUpload\Exceptions\UploadMissingFileException;
-use Pion\Laravel\ChunkUpload\Receiver\FileReceiver;
-
-class ChunkFileUploadController extends Controller
-{
-    /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->middleware('auth');
-    }
-
-    public function uploadFile(FileReceiver $receiver)
-    {
-        // check if the upload is success, throw exception or return response you need
-        if ($receiver->isUploaded() === false) {
-            throw new UploadMissingFileException();
-        }
-        // receive the file
-        $save = $receiver->receive();
-
-        // check if the upload has finished (in chunk mode it will send smaller files)
-        if ($save->isFinished()) {
-            // save the file and return any response you need
-            return $this->saveFile($save->getFile());
-        }
-
-        // we are in chunk mode, lets send the current progress
-        $handler = $save->handler();
-        return response()->json([
-            "done" => $handler->getPercentageDone()
-        ]);
-    }
-
-    protected function saveFile(UploadedFile $originalFile)
-    {
-        $validator = Validator::make([
-            'file' => $originalFile
-        ], [
-            'file' => [new Extension(['gcode', 'stl'])]
-        ]);
-
-        if($validator->fails()) {
-            return response()->json([
-                "error" => "Unsupported file type: {$originalFile->getClientOriginalExtension()}"
-            ], 422);
-        }
-
-        $file = File::fromUploadedFile($originalFile, Auth::user());
-
-        return redirect()->route('jobs.create.file', [$file]);
-    }
-
-    /**
-     * Create unique filename for uploaded file
-     * @param UploadedFile $file
-     * @return string
-     */
-    protected function createFilename(UploadedFile $file)
-    {
-        $extension = $file->getClientOriginalExtension();
-        $filename = str_replace("." . $extension, "", $file->getClientOriginalName()); // Filename without extension
-
-        // Add timestamp hash to name of the file
-        $filename .= "_" . md5(time()) . "." . $extension;
-
-        return $filename;
-    }
-}
